+++ wazo_chatd/plugins/connectors/plugin.py
@@ -10,6 +10,7 @@ from wazo_auth_client import Client as AuthClient
 from wazo_chatd.plugin_helpers.dependencies import PluginDependencies
 from wazo_chatd.plugins.connectors.bus_consume import BusEventHandler
 from wazo_chatd.plugins.connectors.http import (
+    ConnectorIdentitiesResource,
     ConnectorListResource,
     ConnectorWebhookResource,
     IdentityItemResource,
@@ -67,6 +68,11 @@ class Plugin:
         api.add_resource(
             ConnectorListResource,
             '/connectors',
+            resource_class_args=[router],
+        )
+        api.add_resource(
+            ConnectorIdentitiesResource,
+            '/connectors/<backend>/identities',
             resource_class_args=[router],
         )
         api.add_resource(